+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## [unversioned] - 
 ### Fixed
 - fix css stitching for new chrome 78 (bug in chrome). [Trello 1206](https://trello.com/c/euVqe1Sv)
+- Default viewportSize `null` value was overriding configuration. [Trello 1168](https://trello.com/c/yPqI3erm)
 
 ## [2.11.1] - 2019-11-5
 ### Added